A major rework of spectrum-css processing
This is a complete rework of the script that processes the spectrum-css into a format that web components can consume.

The original script used regular expressions to do the conversion. It made a great proof of concept, but was becoming increasingly hard to extend or maintain. With this change we now use full selector parsing into ASTs for the purpose of doing the transformations. This allows much more semantically correct handling of the selector parts.

As well, a few new features were added to remove the need for the custom transforms in the button and side-nav components.