fix: factor theme to use a single DOM node
This addresses issue #154.

Currently themes are implemented as a heirarchy of nodes to get the CSS custom property scoping correct. This change crunches that down into a single node.

The PR also renames the `themes` package to `theme` to match the naming converntion in place for the rest of the components.