feat: work in progress on updating sp-menu item ownership
Now "sp-menu-item-added" and "sp-menu-item-removed" events are handled
to pick up the menu items a parent menu should own. To get the right DOM
order we do a debounced update.
Looks like some focus issues got introduced, at least in the Picker
context, and there's likely more issues with this that need to be fixed
for menus with groups that have different ownership