refactor(illustrated-message): remove cta variant
The CTA variant is no longer necessary, as it provides no unique styles.
It has been removed in the spectrum CSS component.

This variant was at a previous point in time differing only in whether
the description text was italic or not. The more recent versions
(including current 0.9.14 illustrated-message SWC) as dictated by design
and tokens, have a normal font-style by default.