refactor(table): styles for checkbox cell and head-cell attribute
Previously all styles for Table-cell, Table-headCell, and
Table-checkboxCell were being applied to the sp-table-checkbox-cell,
regardless of where it was located. Resulting in head cell styles
appearing on checkbox body cells, and other incorrect styles.

Adds a 'headCell' property / 'head-cell' attribute to the checkbox
cell component. And modifies the spectrum-config to include the various
styles depending on whether that attribute exists or not.