docs(table): clarify docs around current use of scroller
Scrolling in previous examples that were not virtualized was a result
of the example wrapper's overflow and not a feature of the table. So
setting a height on these was only triggering scrolling in that wrapper.

There was also a non-virtualized example with 'scroller' set to true
that wasn't doing anything (that feature only appears to work with
virtualized examples).