fix(textfield,numberfield): inline sizing
Notes: the main thing this commit does is allow the field label's
inline-size to be set to auto, so that its width is independent of the
input.

Unfortunately this also affects help text, whose previous behavior was
to align its width to the input.

If the goal is to pass VRTs, we'll need to adjust the help text to align
its width to the input.

This behavior (label width independent of input, helptext width aligned
with input) feels odd.