fix: provide field label modifier
Because now the field label is inside of sp-textfield, its width is
dependent on the width of sp-textfield. Figma/CSS/React/Design
Guidelines all indicate that this is the correct pattern. In case this
has a negative effect on any consumer who relied on the textfield and
label having indpendent widths, a mod custom property has been provided
to override this behavior.